final-pjt-back/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
Django==3.2.18
django-cors-headers==4.0.0
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.0
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
vue==0.0.1
wcwidth==0.2.6
zipp==3.15.0
